+++ CHANGELOG.md
@@ -1,7 +1,13 @@
-# UNRELEASED
+# CHANGELOG
 
-# 0.1.2 (January 6th, 2022)
+## UNRELEASED
 
-FEATURES
+## RELEASED
 
+### 0.2.0 (Jan 30th, 2024)
 
+- Add `get_ref`, `get_mut` and `into_components` APIs
+
+### 0.1.0 (2024)
+
+- `AsyncPeekable` and `Peekable`